requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
fastapi==0.115.12
griffe==1.7.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
jiter==0.9.0
MarkupSafe==3.0.2
mcp==1.6.0
openai==1.76.0
-e git+https://github.com/openai/openai-agents-python.git@4187fba9556683c5ace4706ebd5dbe90cd14e315#egg=openai_agents
pyairtable==3.1.1
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
sse-starlette==2.3.3
starlette==0.46.2
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
